backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
supabase
email-validator
groq
google-genai
moviepy
opencv-python
python-multipart